requirements.txt
Django==4.2.7
python-decouple==3.8
requests==2.31.0
pillow==10.4.0
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
python-dateutil==2.8.2